refactor: move color_setters macro to a separate file and update NumberedListItemBlock implementation
