+++ notionrs/src/macro/mod.rs
@@ -1,2 +1,3 @@
+mod color_setters;
 mod impl_display_from_string_field;
 mod impl_from_as_ref;
